refactor context menu of graphic views
